% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{\%_>=_\%}
\alias{\%_>=_\%}
\title{"Greater Than Or Equal To"}
\usage{
lhs \%_>=_\% rhs
}
\arguments{
\item{lhs}{Left-hand side of operator.}

\item{rhs}{Right-hand side of operator.}
}
\description{
Special NA >= Operator.
}
\examples{
2 \%_>=_\% 1
1 \%_>=_\% NA
NA \%_>=_\% NA
}
