% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{\%_<=_\%}
\alias{\%_<=_\%}
\title{"Less Than Or Equal To"}
\usage{
lhs \%_<=_\% rhs
}
\arguments{
\item{lhs}{Left-hand side of operator.}

\item{rhs}{Right-hand side of operator.}
}
\description{
Special NA <= Operator.
}
\examples{
1 \%_<=_\% 2
NA \%_<=_\% 1
NA \%_<=_\% NA
}
