% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%!in\%}
\title{"Not In"}
\usage{
x \%notin\% y

x \%!in\% y
}
\arguments{
\item{x}{Value(s) to be checked for in y.}

\item{y}{Value(s) to be checked against.}
}
\description{
Negation of the \code{\%in\%} operator.
}
\examples{
1 \%notin\% c(1,2,3)
1 \%notin\% c(2,3)
1 \%!in\% c(1,2,3)
1 \%!in\% c(2,3)
}
