% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{quantile_sas}
\alias{quantile_sas}
\title{"Derive Quantiles Like SAS"}
\usage{
quantile_sas(x, probs, na.rm = TRUE)
}
\arguments{
\item{x}{Same as x parameter for stats::quantile.}

\item{probs}{Same as probs parameter for stats::quantile.}

\item{na.rm}{Same as na.rm parameter for stats::quantile.}
}
\description{
"Derive Quantiles Like SAS"
}
\examples{
quantile_sas(stats::runif(10), probs = 0.5)
}
