% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{round_sas}
\alias{round_sas}
\title{"SAS Rounding in R"}
\usage{
round_sas(x, digits = 0)
}
\arguments{
\item{x}{Numeric variable / value to round.}

\item{digits}{Number of digits to display to the right of the decimal place.}
}
\description{
Round in R Like SAS: "Banker's Rounding"
}
\examples{
round_sas(2.5)
round_sas(2.55, digits = 1)
}
