% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasbridg2r.R
\name{round_sas_pct}
\alias{round_sas_pct}
\title{"Display Rounded Percentages Like SAS"}
\usage{
round_sas_pct(x, digits = 0, by100 = TRUE)
}
\arguments{
\item{x}{Numeric variable / value to round.}

\item{digits}{Number of digits to display to the right of the decimal place.}

\item{by100}{Multiply the numeric variable / value by 100.}
}
\description{
Round percentages like SAS and display the \% sign.
}
\examples{
round_sas_pct(0.55)
round_sas_pct(0.555, digits = 1)
round_sas_pct(55.5, by100 = FALSE)
}
